+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- TextField.OnSubmitted event
+
 ## [0.6.0] - 2026-04-23
 ### Added
 - DropdownField